Switch back to JavaScript
Limit to 3 images on potentially metered connections.
