+++ .github/workflows/ci.yml
@@ -10,16 +10,13 @@ jobs:
 
     steps:
       - name: Install dependencies
-        run: apk add gcc imagemagick-dev make musl-dev pkgconfig vips-dev
+        run: apk add gcc imagemagick-dev make musl-dev pkgconfig
 
       - name: Check out the repo
         uses: actions/checkout@v4
 
       - name: Build
-        run: make server img-out
-
-      - name: Build
-        run: make image-resizer
+        run: make imgen
 
       - name: Run tests
         run: go test ./...
@@ -31,7 +28,7 @@ jobs:
 
     steps:
       - name: Install dependencies
-        run: apk add gcc imagemagick-dev musl-dev pkgconfig tar vips-dev
+        run: apk add gcc imagemagick-dev musl-dev pkgconfig tar
 
       - name: Check out the repo
         uses: actions/checkout@v4
@@ -43,23 +40,3 @@ jobs:
         uses: golangci/golangci-lint-action@v7
         with:
           install-mode: none
-
-#  test-js:
-#    name: Build and test Javascript code
-#    runs-on: ubuntu-latest
-#
-#    steps:
-#      - name: Check out code into the Go module directory
-#        uses: actions/checkout@v2
-#
-#      - uses: actions/setup-node@v2
-#        with:
-#          node-version: '17'
-#          cache: 'npm'
-#
-#      - run: npm install
-#
-#      - run: npm test
-#
-#      - name: Build the webapp
-#        run: make webapp
