+++ Makefile
@@ -1,19 +1,16 @@
-all: backgrounds webapp
+FONTS := $(addprefix web-src/webfonts/,fa-brands.woff2 fa-solid.woff2)
 
-imgen: $(shell find . -name '*.go' -type f) go.mod go.sum
-	go build -o $@ ./cmd/imgen
-
-backgrounds: imgen
-	mkdir -p $@
-	./$< -out-dir $@ -in-dir img-src
-
-fontawesome-subsets:
-	make -C fa-src
-	mv fa-src/fa-brands.woff2 fa-src/fa-solid.woff2 web-src/webfonts/
-
-webapp: fontawesome-subsets backgrounds
+all: $(FONTS) backgrounds/backgrounds.json
+	mkdir -p dist
 	npx webpack --mode production
 	cp -r backgrounds dist/backgrounds
 
+$(FONTS):
+	make -C fa-src
+	mv fa-src/fa-brands.woff2 fa-src/fa-solid.woff2 web-src/webfonts/
+
+.PHONY: clean
 clean:
 	rm -fr imgen img-out
+
+include backgrounds.mk
