+++ web-src/css/style.css
@@ -505,8 +505,12 @@ code {
   padding-left: 0;
 }
 
-#bg div {
-  background-position: center;
+#bg div img {
+  object-fit: cover;
+  width: 100%;
+  height: 100%;
+  vertical-align: middle;
+  justify-content: center;
 }
 
 @media (orientation: portrait) {
