+++ webpack.config.js
@@ -1,12 +1,13 @@
 const path = require('path');
-
 const CompressionPlugin = require("compression-webpack-plugin");
 const FaviconsWebpackPlugin = require('favicons-webpack-plugin')
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const zlib = require("zlib");
 
+const DIST_DIR_NAME = 'dist'
+
 module.exports = {
-    entry: path.resolve(__dirname, 'web-src/app.ts'),
+    entry: path.resolve(__dirname, 'web-src/app.js'),
     plugins: [
         new CompressionPlugin(), // gzip by default
         new CompressionPlugin({
@@ -38,16 +39,15 @@ module.exports = {
             {
                 test: /\.css$/i,
                 use: ['style-loader', 'css-loader'],
-            },
-            {
-                test: /\.ts$/,
-                use: 'ts-loader',
-                exclude: /node_modules/,
-            },
+            }
         ]
     },
     output: {
+        clean: { keep: /backgrounds/},
         filename: '[name].[contenthash].js',
-        path: path.resolve(__dirname, 'dist'),
+        path: path.resolve(__dirname, DIST_DIR_NAME),
     },
+    optimization: {
+        minimize: true,
+    }
 };
